Add role-user-operator to default cards when server is run
Change-type: minor
